[Category] moving Tag request to the request folder
