+++ 04-delivery/src/main/scala/com/khanr1/cryocompose/controllers/inMemory/TagController.scala
@@ -100,19 +100,3 @@ object TagController:
             .asJson
             .pipe(Ok(_))
         }
-
-object request:
-  object Tag:
-    case class Create(name: TagName)
-
-    object Create:
-      def create(name: String): ValidatedNel[String, Create] =
-        TagName.either(name).map(Create(_)).toValidatedNel
-      given Decoder[Create] = Decoder
-        .forProduct1("name")(create(_))
-        .emap(
-          _.toEither.leftMap(x => x.mkString_("\n"))
-        )
-      given entityDecoder[F[_]: cats.effect.Concurrent]: EntityDecoder[F, Create] = jsonOf
-      given Encoder[Create] = Encoder.forProduct1("name")(n => n.name)
-      given entityEncoder[F[_]: cats.effect.Concurrent]: EntityEncoder[F, Create] = jsonEncoderOf
